pyproject.toml
[project]
name = "gepa-mutations"
version = "0.1.0"
description = "Experimental framework for testing and mutating the GEPA prompt evolution framework"
requires-python = ">=3.12,<3.15"
dependencies = [
    "gepa[full]",
    "dspy",
    "boto3",
    "pandas",
    "numpy",
    "matplotlib",
    "plotly",
    "typer",
    "rich",
    "python-telegram-bot",
    "pydantic-settings",
]

[project.scripts]
gepa-mutations = "gepa_mutations.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gepa_mutations"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.uv]
# Exclude gepa/ from workspace discovery so it's treated as a path dependency, not a workspace member
workspace = { exclude = ["gepa"] }

[tool.uv.sources]
gepa = { path = "./gepa", editable = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
